+++ test/ControlServer.js
@@ -1,18 +1,17 @@
 
 const _ = require('lodash');
 const assert = require('chai').assert;
-const rpc = require('jrpc2');
-const Promise = require('bluebird');
 const nconf = require('nconf');
+const getPort = require('get-port');
 
-const logger = require('../src/winston-silent-logger');
-
+nconf.use('memory');
 require(`${__dirname}/../src/nconf_load_env.js`)(nconf);		
 nconf.defaults(require(`${__dirname}/../src/default_config.js`));
-const { ControlServer } = require('../');
+const { ControlServer, TorPool, HTTPServer, SOCKSServer, DNSServer } = require('../');
 
-let controlServer = new ControlServer(logger, nconf);
+let controlServer = new ControlServer(null, nconf);
 let controlPort;
+
 describe('ControlServer', function () {
 	describe('#listen(port)', function () {
 		it('should bind to a given port', async function () {
@@ -24,7 +23,7 @@ describe('ControlServer', function () {
 		it('should create a TorPool with a given configuration', function () {
 			let torPool = controlServer.createTorPool({ ProtocolWarnings: 1 });
 
-			assert.ok((controlServer.torPool instanceof (TorRouter.TorPool)));
+			assert.instanceOf(controlServer.torPool, TorPool);
 			assert.equal(1, torPool.default_tor_config.ProtocolWarnings);
 		});
 	});
@@ -60,5 +59,3 @@ describe('ControlServer', function () {
 		await controlServer.torPool.exit();
 	});
 });
-
-require('./RPCServer');