+++ test/DNSServer.js
@@ -3,15 +3,18 @@ const nconf = require('nconf');
 const getPort = require('get-port');
 const dns = require('native-dns');
 
-const logger = require('../src/winston-silent-logger');
 const { TorPool, DNSServer } = require('../');
 const { WAIT_FOR_CREATE } = require('./constants');
+
+nconf.use('memory');
+require(`${__dirname}/../src/nconf_load_env.js`)(nconf);		
+nconf.defaults(require(`${__dirname}/../src/default_config.js`));
 
 let dnsServerTorPool;
 let dnsServer;
 describe('DNSServer', function () {
-	dnsServerTorPool = new TorPool(nconf.get('torPath'), {}, nconf.get('parentDataDirectory'), 'round_robin', null, logger);
-	dnsServer = new DNSServer(dnsServerTorPool, {}, 10000, logger);
+	dnsServerTorPool = new TorPool(nconf.get('torPath'), {}, nconf.get('parentDataDirectory'), 'round_robin', null);
+	dnsServer = new DNSServer(dnsServerTorPool, {}, 10000);
 	let dnsPort;
 	before('start up server', async function (){
 		this.timeout(WAIT_FOR_CREATE);
@@ -19,7 +22,7 @@ describe('DNSServer', function () {
 		await dnsServerTorPool.create(1);
 		dnsPort = await getPort();
 
-		dnsServer.serve(dnsPort);
+		await dnsServer.listen(dnsPort);
 	});
 
 	describe('#handle_dns_request(req, res)', function () {
@@ -32,7 +35,7 @@ describe('DNSServer', function () {
 					type: 'A'
 				}),
 				server: { address: '127.0.0.1', port: dnsPort, type: 'udp' },
-				timeout: 1000,
+				timeout: 10000,
 			});
 
 			req.on('timeout', function () {